# docs/ads/openapi.yaml
openapi: 3.0.3
info:
  title: GA4GH Access Decision Service (ADS)
  version: 1.0.0
  description: |
    ADS bridges GA4GH AAI Passports and resource services by managing access requests,
    DUO evaluation, DAC workflows, grants, visa export, and authorization introspection.
servers:
  - url: https://ads.example.org/ads/v1
paths:
  /researchers/{id}:
    get:
      summary: Get researcher profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Researcher record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Researcher"
  /researchers/sync:
    post:
      summary: Sync researcher profile and apply institutional mappings
      description: Called by the AAI broker after upstream login (DAC API key).
      security:
        - dacApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearcherSyncRequest"
      responses:
        "200":
          description: Grants created from institutional permission mappings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grant"
  /researchers/{id}/signed-visas:
    get:
      summary: Fetch signed visa JWTs for passport assembly
      security:
        - dacApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signed visa JWTs from visa-registry via ADS export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedVisasResponse"
  /researchers/{id}/visas:
    get:
      summary: Export visa claims for AAI passport assembly
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unsigned visa claims derived from grants and affiliations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearcherVisasResponse"
  /datasets:
    post:
      summary: Register a dataset
      security:
        - dacApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetRequest"
      responses:
        "200":
          description: Created dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/{id}:
    get:
      summary: Get dataset by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /projects:
    post:
      summary: Register a research project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Created project
  /projects/{id}:
    get:
      summary: Get research project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Research project
  /duo/evaluate:
    post:
      summary: Evaluate DUO compatibility
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuoEvaluateRequest"
      responses:
        "200":
          description: DUO evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuoEvaluationResult"
  /access-requests:
    post:
      summary: Submit an access request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccessRequestBody"
      responses:
        "200":
          description: Created access request (may be auto-approved)
  /access-requests/{id}:
    get:
      summary: Get access request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Access request
  /dac/requests:
    get:
      summary: List pending DAC queue
      security:
        - dacApiKey: []
      responses:
        "200":
          description: Pending and escalated requests
  /dac/requests/{id}/approve:
    post:
      summary: Approve access request
      security:
        - dacApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DacActionRequest"
      responses:
        "200":
          description: Approved request and created grant
  /dac/requests/{id}/reject:
    post:
      summary: Reject access request
      security:
        - dacApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rejected request
  /dac/requests/{id}/escalate:
    post:
      summary: Escalate access request
      security:
        - dacApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Escalated request
  /grants:
    get:
      summary: List grants
      security:
        - bearerAuth: []
        - dacApiKey: []
      parameters:
        - name: researcher_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Active grants
  /grants/{id}:
    get:
      summary: Get grant
      security:
        - bearerAuth: []
        - dacApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Grant
    delete:
      summary: Revoke grant
      security:
        - dacApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Revoked grant
  /introspect:
    post:
      summary: Authorize a Passport for a resource
      description: |
        Resource services (Beacon, DRS, htsget, WES, TES) call this endpoint with a
        researcher Passport JWT and resource identifier.
      security:
        - dacApiKey: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntrospectRequest"
      responses:
        "200":
          description: Introspection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrospectResponse"
  /permission-sources:
    post:
      summary: Register institutional OIDC permission source
      security:
        - dacApiKey: []
      responses:
        "200":
          description: Permission source
  /permission-mappings:
    post:
      summary: Map OIDC claim value to dataset grant
      security:
        - dacApiKey: []
      responses:
        "200":
          description: Permission mapping
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    dacApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Researcher:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        email:
          type: string
        affiliations:
          type: array
          items:
            $ref: "#/components/schemas/ResearcherAffiliation"
    ResearcherAffiliation:
      type: object
      properties:
        organization:
          type: string
        role:
          type: string
    ResearcherVisasResponse:
      type: object
      properties:
        researcher_id:
          type: string
        visas:
          type: array
          items:
            $ref: "#/components/schemas/VisaClaim"
    VisaClaim:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
        source:
          type: string
        asserted:
          type: integer
    CreateDatasetRequest:
      type: object
      required: [name, duo_codes]
      properties:
        name:
          type: string
        description:
          type: string
        duo_codes:
          type: array
          items:
            type: string
        external_id:
          type: string
        auto_approve_enabled:
          type: boolean
        auto_approve_threshold:
          type: integer
          minimum: 0
          maximum: 100
    Dataset:
      allOf:
        - $ref: "#/components/schemas/CreateDatasetRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
    CreateProjectRequest:
      type: object
      required: [researcher_id, name, duo_codes]
      properties:
        researcher_id:
          type: string
        name:
          type: string
        duo_codes:
          type: array
          items:
            type: string
    DuoEvaluateRequest:
      type: object
      properties:
        dataset_duo:
          type: array
          items:
            type: string
        dataset_id:
          type: string
          format: uuid
        project_duo:
          type: array
          items:
            type: string
        project_id:
          type: string
          format: uuid
        auto_approve_threshold:
          type: integer
    DuoEvaluationResult:
      type: object
      properties:
        compatible:
          type: boolean
        score:
          type: integer
        auto_approvable:
          type: boolean
        reason:
          type: string
        matched_codes:
          type: array
          items:
            type: string
        missing_codes:
          type: array
          items:
            type: string
        procedural_modifiers:
          type: array
          items:
            type: string
    CreateAccessRequestBody:
      type: object
      required: [researcher_id, dataset_id, project_id]
      properties:
        researcher_id:
          type: string
        dataset_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        justification:
          type: string
    DacActionRequest:
      type: object
      properties:
        reason:
          type: string
        actor:
          type: string
    IntrospectRequest:
      type: object
      required: [token, resource]
      properties:
        token:
          type: string
          description: GA4GH Passport JWT
        resource:
          type: string
        action:
          type: string
        dataset_id:
          type: string
          format: uuid
    IntrospectResponse:
      type: object
      properties:
        active:
          type: boolean
        sub:
          type: string
        grant_ids:
          type: array
          items:
            type: string
            format: uuid
        duo_codes:
          type: array
          items:
            type: string
        exp:
          type: integer
        reason:
          type: string
    Grant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        researcher_id:
          type: string
        dataset_id:
          type: string
          format: uuid
        source:
          type: string
    ResearcherSyncRequest:
      type: object
      required: [sub]
      properties:
        sub:
          type: string
        display_name:
          type: string
        email:
          type: string
        claims:
          type: object
          additionalProperties: true
        affiliations:
          type: array
          items:
            type: object
            properties:
              organization:
                type: string
              role:
                type: string
    SignedVisasResponse:
      type: object
      properties:
        researcher_id:
          type: string
        visa_jwts:
          type: array
          items:
            type: string